refactor iterator to use tsitertorstate
